requirements.txt
asciimatics>=1.15.0
PyYAML>=6.0
